+++ react-components/src/architecture/base/reactUpdaters/DomainObjectPanelUpdater.ts
